api/requirements.txt
fastapi
uvicorn
pyspark
delta-spark
matplotlib
pandas>=2.0.0
